fix CA1506
